Add ATTR_FREQ_PAU_VPD_MHZ attribute
This is the FREQ attribute that get setup by Hostboot HWPs but feed into the
sbe scratch regs. It should be preserved across IPLs. Updated the necessary
files to create that attribute.

Output from the generated dts file.

ATTR_FREQ_PAU_MHZ = <2133>;
ATTR_FREQ_PAU_VPD_MHZ = <0>;
ATTR_FREQ_PCIE_MHZ = <2000>;

From the above lines it is clear that it is created similar to other frequency
attributes and the value is set to 0. Also verified that no other attributes
are affected by comparing the difference between the dts file generated with
and without the changes.

Signed-off-by: deepakala-k <deepakala.karthikeyan@ibm.com>