Add new attribute inventory index
When the PLDM requests for a SBE dump, it sends the inventory index of the
dimm. Using that index, we need to find the corresponding OCMB on the device
tree. As the index attribute does not have corelation to inventory index, it is
difficult to map the inventory item to the device tree target. Introduced the
new attribute that will help in mapping the inventory item to the pdbg target

This commit reverts some of the changes that were added in the previous commit
153dde5330b0b998b0705cb3e5eba40e1a53b2cc that modified the existing index that
broke the ecmdquery command making it mismatch with the cronus command

Tested and the new attribute ATTR_INVENTORY_INDEX is created as expected.

Signed-off-by: deepakala-k <deepakala.karthikeyan@ibm.com>